requirements.txt
anyio==3.7.1
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.2.0
exceptiongroup==1.1.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
numpy==1.25.1
pandas==2.0.3
python-dateutil==2.8.2
python-telegram-bot==20.4
pytz==2023.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
tabulate==0.9.0
tzdata==2023.3
urllib3==2.0.3
